Move preprocessor-plugin into src/ with other JS
