+++ src/preprocessor-plugin.js
